Remove postcss-custom-properties
