+++ package.json
@@ -16,15 +16,15 @@
     "start:build": "npm-run-all -p 'build:* -- --serve --watch'",
     "start:static": "onchange -k 'src/static/**' -- npm run build:static",
     "start:styles": "npm-run-all -p 'start:styles:*'",
-    "start:styles:main": "onchange -k 'postcss/main/*' 'src/styles/main/**' 'src/components/**/*.css' -- npm run build:styles:main",
-    "start:styles:variables": "onchange -k 'postcss/variables/*' 'src/styles/variables/**' 'src/design/**' -- npm run build:styles:variables",
+    "start:styles:main": "onchange -k 'postcss.config.js' 'src/styles/main/**' 'src/components/**/*.css' -- npm run build:styles:main",
+    "start:styles:variables": "onchange -k 'src/styles/variables/**' 'src/design/**' -- npm run build:styles:variables",
     "start:bootstrap": "onchange -k 'src/bootstrap/**' 'tasks/generate-bootstrap.js' -- npm run build:bootstrap",
     "build": "npm run clean && NODE_ENV=production npm-run-all -p 'build:*'",
     "build:static": "cp -r src/static/* dist && cp node_modules/bootstrap/dist/js/bootstrap.bundle.min.js* dist/scripts/bootstrap && :",
     "build:patternlib": "uiengine build",
     "build:styles": "npm-run-all -s 'build:styles:*' && :",
-    "build:styles:main": "postcss --config postcss/main --output dist/styles/btcpayserver-main.css src/styles/main/index.css",
-    "build:styles:variables": "echo '' | postcss --config postcss/variables && postcss --config postcss/main --output generated/theme.css src/styles/variables/theme.css && sed -i 's/\t/  /g' generated/design.css && cat generated/*.css | sed -z 's/\\n}\\n:root {//' > dist/styles/btcpayserver-variables.css",
+    "build:styles:main": "postcss --output dist/styles/btcpayserver-main.css src/styles/main/index.css",
+    "build:styles:variables": "echo '' | npm run generate-design && postcss --output generated/theme.css src/styles/variables/theme.css && sed -i 's/\t/  /g' generated/design.css && cat generated/*.css | sed -z 's/\\n}\\n:root {//' > dist/styles/btcpayserver-variables.css",
     "build:bootstrap": "npm run generate-bootstrap && :",
     "optimize": "npm-run-all -p optimize:*",
     "optimize:styles": "npm-run-all -p optimize:styles:*",
@@ -34,6 +34,7 @@
     "prod": "npm-run-all build optimize",
     "generate-color-palette": "node ./tasks/generate-color-palette.js",
     "generate-bootstrap": "node ./tasks/generate-bootstrap.js",
+    "generate-design": "node ./tasks/generate-design.js",
     "copy-assets": "bash ./tasks/copy-assets-to-targets.sh",
     "build-and-copy": "npm-run-all -s clean build build:static copy-assets"
   },
@@ -50,7 +51,6 @@
     "onchange": "7.1.0",
     "postcss": "8.4.19",
     "postcss-cli": "10.0.0",
-    "postcss-custom-properties": "12.1.9",
     "postcss-import": "15.0.0",
     "postcss-mixins": "9.0.4",
     "postcss-nesting": "10.2.0",
